refactor: obtain the information for each rank and then parse the relevent information
Before, it would go through the entire page twice, to find the title and artist of each rank.
Now, it will go through the page once, get all the ranks, then go through the ranks to get the rank, title, and artist.
Still need to find a way to obtain length of each songs. (Could possibly use spotipy and create a playlist with every song)